sstables: storage: define opened_directory in header file
So it can be used outside the storage module
in the following patches.
