+++ sstables/storage.cc
@@ -38,39 +38,6 @@
 
 namespace sstables {
 
-class opened_directory final {
-    std::filesystem::path _pathname;
-    file _file;
-
-public:
-    explicit opened_directory(std::filesystem::path pathname) : _pathname(std::move(pathname)) {};
-    explicit opened_directory(const sstring &dir) : _pathname(std::string_view(dir)) {};
-    opened_directory(const opened_directory&) = delete;
-    opened_directory& operator=(const opened_directory&) = delete;
-    opened_directory(opened_directory&&) = default;
-    opened_directory& operator=(opened_directory&&) = default;
-    ~opened_directory() = default;
-
-    const std::filesystem::path::string_type& native() const noexcept {
-        return _pathname.native();
-    }
-
-    const std::filesystem::path& path() const noexcept {
-        return _pathname;
-    }
-
-    future<> sync(io_error_handler error_handler) {
-        if (!_file) {
-            _file = co_await do_io_check(error_handler, open_directory, _pathname.native());
-        }
-        co_await do_io_check(error_handler, std::mem_fn(&file::flush), _file);
-    };
-
-    future<> close() {
-        return _file ? _file.close() : make_ready_future<>();
-    }
-};
-
 // cannot define these classes in an anonymous namespace, as we need to
 // declare these storage classes as "friend" of class sstable
 class filesystem_storage final : public sstables::storage {
